STRIPES-827 provide stripes v7 compatibility
Indicate compatibility with `@folio/stripes` `v7`. The breaking changes
from `v6` (react 17, rxjs 6) are not significant here, thus this is not
a breaking change. It's possible to argue that adding the `react-dom`
peer-dep and a version to the `react` peer-dep mean this _is_ a breaking
change, and you'd probably win that fight. Practically speaking,
however, those were always requirements but were just not correctly
stated. :shrug:

Refs STRIPES-827