STRIPES-832 use alpha-prefixes for HTML id attributes
Using a naked UUID as an html `id` attribute will cause problems when
trying to retrieve elements using CSS selectors because they can start
with 0-9 and a-f but a CSS selector can't start with a number. IOW, even
though an HTML `id` is technically allowed to start with and contain any
character, in order to play nice with CSS selectors they need to start
with a letter. And now these two do.

Fixes the bug introduced by #19.

Refs STRIPES-832