STRIPES-908 add Barcode to allow-list
JSC for `<Barcode>` is handled specifically here, and so needs to be
specifically escaped because otherwise the sanitizer will see it as an
unknown string and stripe it.

This is a follow-up to #58 where we added sanitization but not escaping,
which had the effect of eliminating XSS (yay!) ... but also eliminating
barcodes (whoops).

Refs STRIPES-908