STRIPES-908 add Barcode to allow-list (#61)
JSX for `<Barcode>` is handled specifically here, and so needs to be
specifically escaped because otherwise the sanitizer will see it as an
unknown tag and strip it.

This is a follow-up to #58 where we added sanitization but not escaping,
which had the effect of eliminating XSS (yay!) ... but also eliminating
barcodes (whoops).

Refs STRIPES-908

(cherry picked from commit dc4d1db8a5f5688c521317170593bd4f1e320ad3)